requirements.txt
Django==5.1.2
django-ninja==1.3.0
pillow==11.0.0
python-decouple==3.8
